requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.4.3
aiohttp==3.11.18
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.6.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==3.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
Babel==2.12.1
beautifulsoup4==4.12.3
billiard==4.2.0
black==23.10.1
blinker==1.8.2
celery==5.3.4
certifi==2023.11.17
cffi==1.17.1
channels==4.2.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comtypes==1.4.8
constantly==23.10.4
crispy-bootstrap5==2023.10
cryptography==44.0.2
daphne==4.1.2
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
Django==5.2
django-autoslug==1.9.9
django-crispy-forms==2.1
django-debug-toolbar==4.2.0
django-eventstream==5.3.2
django-extensions==3.2.3
django-filter==24.3
django-grip==3.5.2
django-modelcluster==6.4
django-permissionedforms==0.1
django-stubs-ext==5.2.0
django-taggit==6.1.0
django-tasks==0.7.0
django-treebeard==4.7.1
django-unfold==0.48.0
django-utils-six==2.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.0
draftjs_exporter==5.1.0
drf-spectacular==0.26.5
drf-yasg==1.21.8
easy-thumbnails==2.10
et_xmlfile==2.0.0
executing==2.0.1
fastapi==0.115.3
filelock==3.16.1
filetype==1.2.0
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
frozenlist==1.5.0
fsspec==2024.10.0
google==3.0.0
greenlet==3.1.1
gripcontrol==4.2.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.26.2
hyperlink==21.0.0
idna==3.6
incremental==24.7.2
inflection==0.5.1
ipython==8.19.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
kombu==5.3.3
laces==0.1.2
magic-filter==1.0.12
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib-inline==0.1.6
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
mysql==0.0.3
mysql-connector-python==9.2.0
mysqlclient==2.2.7
numpy==2.1.3
openai==0.28.0
opencv-python==4.10.0.84
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==23.2
parso==0.8.3
pathspec==0.11.2
pexpect==4.9.0
Pillow==10.1.0
pillow_heif==0.21.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
propcache==0.2.0
ptyprocess==0.7.0
pubcontrol==3.5.0
pure-eval==0.2.2
pushpin==0.5.0
pyaes==1.6.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.17.2
PyJWT==2.8.0
pyOpenSSL==25.0.0
pypiwin32==223
PySocks==1.7.1
pyTelegramBotAPI==4.26.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pyttsx3==2.98
pytz==2023.3.post1
pywin32==308
PyYAML==6.0.1
redis==5.0.1
referencing==0.30.2
regex==2024.11.6
requests==2.31.0
rich==13.8.1
rpds-py==0.12.0
rsa==4.9.1
safetensors==0.4.5
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.31.0
service-identity==24.2.0
setuptools==75.8.0
six==1.16.0
slugify==0.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SpeechRecognition==3.11.0
SQLAlchemy==2.0.35
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.41.0
telebot==0.0.5
telepath==0.3.1
Telethon==1.39.0
text-unidecode==1.3
threadpoolctl==3.5.0
tokenizers==0.21.1
tomli==2.0.1
tqdm==4.66.5
traitlets==5.14.1
transformers==4.50.3
trio==0.29.0
trio-websocket==0.12.2
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.9
websocket-client==1.8.0
Werkzeug==3.0.4
Willow==1.10.0
wsproto==1.2.0
yarl==1.17.1
zope.interface==7.2
